+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/db/PendingTransactionDb.kt
@@ -70,10 +70,10 @@ interface PendingTransactionDao {
     suspend fun removeRawTransactionId(id: Long)
 
     @Query("UPDATE pending_transactions SET minedHeight = :minedHeight WHERE id = :id")
-    suspend fun updateMinedHeight(id: Long, minedHeight: Int)
+    suspend fun updateMinedHeight(id: Long, minedHeight: Long)
 
     @Query("UPDATE pending_transactions SET raw = :raw, rawTransactionId = :rawTransactionId, expiryHeight = :expiryHeight WHERE id = :id")
-    suspend fun updateEncoding(id: Long, raw: ByteArray, rawTransactionId: ByteArray, expiryHeight: Int?)
+    suspend fun updateEncoding(id: Long, raw: ByteArray, rawTransactionId: ByteArray, expiryHeight: Long?)
 
     @Query("UPDATE pending_transactions SET errorMessage = :errorMessage, errorCode = :errorCode WHERE id = :id")
     suspend fun updateError(id: Long, errorMessage: String?, errorCode: Int?)
